refactor: Generalize internal event emitter
BREAKING CHANGE:
- Trace span event emitter is relocated, and is accessible at `_eventsEmitter` instead of `_traceSpanEmitter` property on SDK
- Trace span events are renamed from `open` to `trace-span-open` and `close` to `trace-span-close`
