+++ node/packages/sdk/lib/emitter.js
